🚸 Move share button to context menu
